oops commited a war crime

+++ dragonship/src/main/java/dev/vinpol/nebula/dragonship/geo/Coordinate.java
@@ -1,4 +1,4 @@
 package dev.vinpol.nebula.dragonship.geo;
 
-public record Coordinate(int x, int y) {
+public record Coordinate(double x, double y) {
 }
